Add a new `proxy` verb
This command runs a webserver that can fetch manifests and blobs.
One intended use case is for ostree+containers integration.  The goal
is to fetch and *stream* the underlying tarball from a container
image.  The ostree code here will do things like validate the embedded
OSTree signatures inside the tarball.

Deisgn constraints:

 - This should not read or write to the containers/image storage; we
   want to stream the result and have other code unpack it
   into the target ostree repo
 - But, we *do* want to reuse any configuration in `/etc/containers/registries.conf`
   particularly around mirroring.
 - We want to reuse the containers/image default handling for
   authentication (`~/.docker/config.json` etc.)
 - This should be runnable fully unprivileged under e.g. a transient
   uid; again should have no effect on the system

It's of course a bonus here if we also support containers signing, etc.

The expected primary use case is that ostree tooling allocates a `socketpair()`
and passes it down to skopeo, which will run a webserver on that anonymous
socketpair.

For ease of debugging/testing, we also support listening on a localhost TCP port.